removed gameLogic related things from gameScreen
